+++ openS3/__init__.py
@@ -2,7 +2,7 @@ from .utils import OpenS3
 
 __author__ = 'Paul Logston'
 __email__ = 'code@logston.me'
-__version__ = '0.1.3'
+__version__ = '0.1.4'
 
 __all__ = ('OpenS3',)