+++ openS3/utils.py
@@ -9,7 +9,9 @@ import re
 import urllib.parse
 from wsgiref.handlers import format_date_time
 
-from .constants import MEDIA_TYPES, ENCODING, AWS_DATETIME_FORMAT
+from .constants import (
+    CONTENT_TYPES, ENCODING, AWS_DATETIME_FORMAT, VALID_MODES, DEFAULT_CONTENT_TYPE
+)
 
 
 def b64_string(byte_string):
@@ -82,37 +84,24 @@ class OpenS3(object):
         self.secret_key = secret_key
         validate_values(validation_func=lambda value: value is not None, dic=locals())
         self.netloc = '{}.s3.amazonaws.com'.format(bucket)
-        self.mode = None
-        self.acl = 'public-read'
+        self.mode = 'rb'
+        self.acl = 'private'
 
         # File like attributes
         self.object_key = ''
         self.buffer = ''
-        self._mimetype = None
+        self._content_type = None
         self.headers = {}
+        self.extra_request_headers = {}
 
-    def __call__(self, object_key, mode='r', mimetype=None, acl='public-read'):
-        """
-        Configure :py:class:`OpenS3` object to write to a specific file in S3.
-
-        :param object_key:
-        :param mode:
-        :param mimetype:
-        """
-        self.mode = mode
-        self.object_key = object_key
-        self.mimetype = mimetype
-        self.acl = acl
-        return self
+    def __call__(self, *args, **kwargs):
+        return self.open(*args, **kwargs)
 
     def __enter__(self):
         return self
 
     def __exit__(self, exc_type, exc_val, exc_tb):
-        if self.mode == 'w' and self.buffer:
-            # TODO Does the old file need to be deleted
-            # TODO from S3 before we write over it?
-            self._put()
+        self.close()
 
     def read(self):
         """
@@ -129,41 +118,88 @@ class OpenS3(object):
 
         :param content:
         """
-        if self.mode != 'w':
-            raise RuntimeError('Must open file in write mode to write to file.')
+        if self.mode not in ('wb', 'ab'):
+            raise RuntimeError('Must open file in write or append mode to write to file.')
         self.buffer = content
         # TODO handle multiple writes to same file.
 
-    @property
-    def mimetype(self):
+    def open(self, object_key,
+             mode='rb', content_type=None, acl='private', extra_request_headers=None):
         """
-        Return mimetype of file. If file does not
-        have a mimetype, make a guess.
-        """
-        if self._mimetype:
-            return self._mimetype
+        Configure :py:class:`OpenS3` object to write to or read from a specific S3 object.
 
-        mimetype = 'binary/octet-stream'
+        :param object_key: A unique identifier for an object a bucket.
+        :param mode: The mode in which the S3 object is opened. See Modes below.
+        :param content_type: A standard MIME type describing the format of the contents.
+        :param acl: Name of a specific canned Access Control List to apply to the object.
+
+        **Modes**
+
+        ====  ===================================================================
+        mode  Description
+        ====  ===================================================================
+        'rb'  open for reading (default)
+        'wb'  open for writing, truncating the file first
+        'ab'  open for writing, appending to the end of the file if it exists
+        ====  ===================================================================
+
+        **Access Control List (acl)**
+
+        Valid values include:
+
+            - private  (*default*)
+            - public-read
+            - public-read-write
+            - authenticated-read
+            - bucket-owner-read
+            - bucket-owner-full-control
+        """
+        if mode not in VALID_MODES:
+            raise ValueError('{} is not a valid mode for opening an S3 object.'.format(mode))
+
+        self.object_key = object_key
+        self.mode = mode
+        self.content_type = content_type
+        self.acl = acl
+        self.extra_request_headers = extra_request_headers if extra_request_headers else {}
+        return self
+
+    def close(self):
+        if self.mode in ('wb', 'ab') and self.buffer:
+            # TODO Does the old file need to be deleted
+            # TODO from S3 before we write over it?
+            self._put()
+
+    @property
+    def content_type(self):
+        """
+        Return content_type of file. If file does not
+        have a content_type, make a guess.
+        """
+        if self._content_type:
+            return self._content_type
+
+        content_type = DEFAULT_CONTENT_TYPE
         # get file extension
         if self.object_key:
             _, extension = os.path.splitext(self.object_key)
             extension = extension.strip('.')
-            if extension in MEDIA_TYPES:
+            if extension in CONTENT_TYPES:
                 # Make an educated guess about what the Content-Type should be.
-                mimetype = MEDIA_TYPES[extension]
+                content_type = CONTENT_TYPES[extension]
 
-        return mimetype
+        return content_type
 
-    @mimetype.setter
-    def mimetype(self, mimetype):
-        self._mimetype = mimetype
+    @content_type.setter
+    def content_type(self, content_type):
+        self._content_type = content_type
 
     @property
     def size(self):
         """
         Return the size of the buffer, in bytes.
         """
-        return len(self.buffer)  # TODO is the right way to get size of buffer (bytes)
+        return len(self.buffer)  # TODO is this the right way to get size of buffer (bytes)
 
     @property
     def url(self):
@@ -196,7 +232,7 @@ class OpenS3(object):
         with closing(HTTPConnection(self.netloc)) as conn:
             conn.request('GET', self.object_key, headers=request_headers)
             response = conn.getresponse()
-            if not response.status in (200,):
+            if response.status not in (200,):
                 if response.length is None:
                     # length == None seems to be returned from GET requests
                     # to non-existing files
@@ -237,7 +273,7 @@ class OpenS3(object):
         with closing(HTTPConnection(self.netloc)) as conn:
             conn.request('DELETE', self.object_key, headers=headers)
             response = conn.getresponse()
-            if not response.status in (204,):
+            if response.status not in (204,):
                 raise S3IOError(
                     'openS3 DELETE error. '
                     'Response status: {}. '
@@ -278,8 +314,11 @@ class OpenS3(object):
         headers['Date'] = format_date_time(datetime.now().timestamp())
         headers['Content-Length'] = self.size
         headers['Content-MD5'] = self.md5hash
-        headers['Content-Type'] = self.mimetype
+        headers['Content-Type'] = self.content_type
         headers['x-amz-acl'] = self.acl
+
+        if self.extra_request_headers:
+            headers.update(self.extra_request_headers)
 
         if self.access_key and self.secret_key:
             string_to_sign_list = [