+++ CHANGELOG.rst
@@ -1,6 +1,14 @@
 Changelog
 =========
 
+0.1.4
+-----
+
+- Changed default ACL value to be *private*.
+- Added :py:meth:`~openS3.utils.OpenS3.open` method.
+- Updated :py:meth:`~openS3.utils.OpenS3.open` to able to be called with a dictionary of extra request headers.
+- Defined specific modes in which an S3 object can be opened.
+
 0.1.3
 -----